Move static class to outside
