+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/storage/TxnMetadata.java
@@ -1,0 +1,35 @@
+/**
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.streamnative.pulsar.handlers.kop.storage;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.Accessors;
+
+/**
+ * TxnMetadata represents the ongoing transaction.
+ */
+@Data
+@Accessors(fluent = true)
+@EqualsAndHashCode
+public class TxnMetadata {
+    private final long producerId;
+    private final long firstOffset;
+    private long lastOffset;
+
+    public TxnMetadata(long producerId, long firstOffset) {
+        this.producerId = producerId;
+        this.firstOffset = firstOffset;
+    }
+}
