+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/storage/ProducerStateManagerTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.kop.coordinator.transaction;
+package io.streamnative.pulsar.handlers.kop.storage;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
@@ -23,8 +23,6 @@ import io.streamnative.pulsar.handlers.kop.KopProtocolHandlerTestBase;
 import io.streamnative.pulsar.handlers.kop.SystemTopicClient;
 import io.streamnative.pulsar.handlers.kop.format.EntryFormatter;
 import io.streamnative.pulsar.handlers.kop.format.EntryFormatterFactory;
-import io.streamnative.pulsar.handlers.kop.storage.PartitionLog;
-import io.streamnative.pulsar.handlers.kop.storage.ProducerStateManager;
 import io.streamnative.pulsar.handlers.kop.systopic.SystemTopicClientFactory;
 import io.streamnative.pulsar.handlers.kop.systopic.SystemTopicProducerStateClient;
 import io.streamnative.pulsar.handlers.kop.utils.timer.MockTime;
@@ -149,13 +147,13 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         appendEndTxnMarker(stateManager, producerId, producerEpoch, ControlRecordType.COMMIT, 27L,
                 -1, time.milliseconds());
 
-        ProducerStateManager.ProducerStateEntry firstEntry = stateManager.lastEntry(producerId).orElseGet(() -> {
+        ProducerStateEntry firstEntry = stateManager.lastEntry(producerId).orElseGet(() -> {
             fail("Expected last entry to be defined");
             return null;
         });
 
-        assertEquals(producerEpoch, firstEntry.getProducerEpoch().shortValue());
-        assertEquals(producerId, firstEntry.getProducerId());
+        assertEquals(producerEpoch, firstEntry.producerEpoch().shortValue());
+        assertEquals(producerId, firstEntry.producerId());
         assertEquals(RecordBatch.NO_SEQUENCE, firstEntry.lastSeq().intValue());
 
         // Fencing should continue to work even if the marker is the only thing left
@@ -180,12 +178,12 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
 
         // The broker should accept the request if the sequence number is reset to 0
         append(stateManager, producerId, producerEpoch, 0, 39L, 4L, false);
-        ProducerStateManager.ProducerStateEntry secondEntry = stateManager.lastEntry(producerId).orElseGet(() -> {
+        ProducerStateEntry secondEntry = stateManager.lastEntry(producerId).orElseGet(() -> {
             fail("Expected last entry to be defined");
             return null;
         });
-        assertEquals(producerEpoch, secondEntry.getProducerEpoch().shortValue());
-        assertEquals(producerId, secondEntry.getProducerId());
+        assertEquals(producerEpoch, secondEntry.producerEpoch().shortValue());
+        assertEquals(producerId, secondEntry.producerId());
         assertEquals(0, secondEntry.lastSeq().intValue());
     }
 
@@ -205,11 +203,11 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         long offset = 735L;
         append(stateManager, producerId, epoch, sequence, offset);
 
-        Optional<ProducerStateManager.ProducerStateEntry> maybeLastEntry = stateManager.lastEntry(producerId);
+        Optional<ProducerStateEntry> maybeLastEntry = stateManager.lastEntry(producerId);
         assertTrue(maybeLastEntry.isPresent());
 
-        ProducerStateManager.ProducerStateEntry lastEntry = maybeLastEntry.get();
-        assertEquals(epoch, lastEntry.getProducerEpoch().shortValue());
+        ProducerStateEntry lastEntry = maybeLastEntry.get();
+        assertEquals(epoch, lastEntry.producerEpoch().shortValue());
         assertEquals(sequence, lastEntry.firstSeq());
         assertEquals(sequence, lastEntry.lastSeq());
     }
@@ -223,12 +221,12 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         appendEndTxnMarker(stateManager, producerId, bumpedProducerEpoch, ControlRecordType.ABORT, 1L,
                 -1, time.milliseconds());
 
-        Optional<ProducerStateManager.ProducerStateEntry> maybeLastEntry = stateManager.lastEntry(producerId);
+        Optional<ProducerStateEntry> maybeLastEntry = stateManager.lastEntry(producerId);
         assertTrue(maybeLastEntry.isPresent());
 
-        ProducerStateManager.ProducerStateEntry lastEntry = maybeLastEntry.get();
-        assertEquals(bumpedProducerEpoch, lastEntry.getProducerEpoch().shortValue());
-        assertEquals(Optional.empty(), lastEntry.getCurrentTxnFirstOffset());
+        ProducerStateEntry lastEntry = maybeLastEntry.get();
+        assertEquals(bumpedProducerEpoch, lastEntry.producerEpoch().shortValue());
+        assertEquals(Optional.empty(), lastEntry.currentTxnFirstOffset());
         assertEquals(RecordBatch.NO_SEQUENCE, lastEntry.firstSeq().intValue());
         assertEquals(RecordBatch.NO_SEQUENCE, lastEntry.lastSeq().intValue());
 
@@ -242,9 +240,9 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         short producerEpoch = 0;
         long offset = 992342L;
         int seq = 0;
-        ProducerStateManager.ProducerAppendInfo producerAppendInfo =
-                new ProducerStateManager.ProducerAppendInfo(
-                        partition.toString(), producerId, ProducerStateManager.ProducerStateEntry.empty(producerId),
+        ProducerAppendInfo producerAppendInfo =
+                new ProducerAppendInfo(
+                        partition.toString(), producerId, ProducerStateEntry.empty(producerId),
                         PartitionLog.AppendOrigin.Client);
 
         producerAppendInfo.appendDataBatch(producerEpoch, seq, seq, time.milliseconds(),
@@ -263,29 +261,29 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         AtomicInteger seq = new AtomicInteger(0);
 
         // Start one transaction in a separate append
-        ProducerStateManager.ProducerAppendInfo firstAppend = stateManager.prepareUpdate(
+        ProducerAppendInfo firstAppend = stateManager.prepareUpdate(
                 producerId, PartitionLog.AppendOrigin.Client);
         appendData(16L, 20L, firstAppend, producerEpoch, seq);
-        assertEquals(new ProducerStateManager.TxnMetadata(producerId, 16L), firstAppend.startedTransactions().get(0));
+        assertEquals(new TxnMetadata(producerId, 16L), firstAppend.startedTransactions().get(0));
         stateManager.update(firstAppend);
         assertEquals(16, stateManager.firstUndecidedOffset().get().longValue());
 
         // Now do a single append which completes the old transaction, mixes in
         // some empty transactions, one non-empty complete transaction, and one
         // incomplete transaction
-        ProducerStateManager.ProducerAppendInfo secondAppend = stateManager.prepareUpdate(
+        ProducerAppendInfo secondAppend = stateManager.prepareUpdate(
                 producerId, PartitionLog.AppendOrigin.Client);
-        Optional<PartitionLog.CompletedTxn> firstCompletedTxn =
+        Optional<CompletedTxn> firstCompletedTxn =
                 appendEndTxn(ControlRecordType.COMMIT, 21L, secondAppend, coordinatorEpoch, producerEpoch);
         assertEquals(
-                Optional.of(new PartitionLog.CompletedTxn(producerId, 16L, 21L, false)),
+                Optional.of(new CompletedTxn(producerId, 16L, 21L, false)),
                 firstCompletedTxn);
         assertEquals(Optional.empty(),
                 appendEndTxn(ControlRecordType.COMMIT, 22L, secondAppend, coordinatorEpoch, producerEpoch));
         assertEquals(Optional.empty(),
                 appendEndTxn(ControlRecordType.ABORT, 23L, secondAppend, coordinatorEpoch, producerEpoch));
         appendData(24L, 27L, secondAppend, producerEpoch, seq);
-        Optional<PartitionLog.CompletedTxn> secondCompletedTxn = appendEndTxn(
+        Optional<CompletedTxn> secondCompletedTxn = appendEndTxn(
                 ControlRecordType.ABORT, 28L, secondAppend, coordinatorEpoch, producerEpoch);
         assertTrue(secondCompletedTxn.isPresent());
         assertEquals(Optional.empty(),
@@ -293,9 +291,9 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         appendData(30L, 31L, secondAppend, producerEpoch, seq);
 
         assertEquals(2, secondAppend.startedTransactions().size());
-        assertEquals(new ProducerStateManager.TxnMetadata(producerId, 24L),
+        assertEquals(new TxnMetadata(producerId, 24L),
                 secondAppend.startedTransactions().get(0));
-        assertEquals(new ProducerStateManager.TxnMetadata(producerId, 30L),
+        assertEquals(new TxnMetadata(producerId, 30L),
                 secondAppend.startedTransactions().get(secondAppend.startedTransactions().size() - 1));
         stateManager.update(secondAppend);
         stateManager.completeTxn(firstCompletedTxn.get());
@@ -303,15 +301,15 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         assertEquals(30L, stateManager.firstUndecidedOffset().get().longValue());
     }
 
-    private Optional<PartitionLog.CompletedTxn> appendEndTxn(
-            ControlRecordType recordType, Long offset, ProducerStateManager.ProducerAppendInfo appendInfo,
+    private Optional<CompletedTxn> appendEndTxn(
+            ControlRecordType recordType, Long offset, ProducerAppendInfo appendInfo,
             short coordinatorEpoch, short producerEpoch) {
         return appendInfo.appendEndTxnMarker(new EndTransactionMarker(recordType, coordinatorEpoch),
                 producerEpoch, offset, time.milliseconds());
     }
 
     public void appendData(Long startOffset, Long endOffset,
-                           ProducerStateManager.ProducerAppendInfo appendInfo,
+                           ProducerAppendInfo appendInfo,
                            short producerEpoch, AtomicInteger seq) {
         int count = (int) (endOffset - startOffset);
         appendInfo.appendDataBatch(producerEpoch, seq.get(), seq.addAndGet(count), time.milliseconds(),
@@ -338,32 +336,32 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         beginTxn(producerId3, producerEpoch, startOffset3);
 
         long lastOffset1 = startOffset3 + 15;
-        PartitionLog.CompletedTxn completedTxn1 =
-                new PartitionLog.CompletedTxn(producerId1, startOffset1, lastOffset1, false);
+        CompletedTxn completedTxn1 =
+                new CompletedTxn(producerId1, startOffset1, lastOffset1, false);
         assertEquals(startOffset2, stateManager.lastStableOffset(completedTxn1));
         stateManager.completeTxn(completedTxn1);
         assertEquals(startOffset2, stateManager.firstUndecidedOffset().get().longValue());
 
         long lastOffset3 = lastOffset1 + 20;
-        PartitionLog.CompletedTxn completedTxn3 =
-                new PartitionLog.CompletedTxn(producerId3, startOffset3, lastOffset3, false);
+        CompletedTxn completedTxn3 =
+                new CompletedTxn(producerId3, startOffset3, lastOffset3, false);
         assertEquals(startOffset2, stateManager.lastStableOffset(completedTxn3));
         stateManager.completeTxn(completedTxn3);
         assertEquals(startOffset2, stateManager.firstUndecidedOffset().get().longValue());
 
         long lastOffset2 = lastOffset3 + 78;
-        PartitionLog.CompletedTxn completedTxn2 =
-                new PartitionLog.CompletedTxn(producerId2, startOffset2, lastOffset2, false);
+        CompletedTxn completedTxn2 =
+                new CompletedTxn(producerId2, startOffset2, lastOffset2, false);
         assertEquals(lastOffset2 + 1, stateManager.lastStableOffset(completedTxn2));
         stateManager.completeTxn(completedTxn2);
         assertEquals(Optional.empty(), stateManager.firstUndecidedOffset());
     }
 
     private void beginTxn(Long producerId, short producerEpoch, Long startOffset) {
-        ProducerStateManager.ProducerAppendInfo producerAppendInfo = new ProducerStateManager.ProducerAppendInfo(
+        ProducerAppendInfo producerAppendInfo = new ProducerAppendInfo(
                 partition.toString(),
                 producerId,
-                ProducerStateManager.ProducerStateEntry.empty(producerId),
+                ProducerStateEntry.empty(producerId),
                 PartitionLog.AppendOrigin.Client
         );
         producerAppendInfo.appendDataBatch(producerEpoch, 0, 0, time.milliseconds(),
@@ -377,7 +375,7 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
     public void testPrepareUpdateDoesNotMutate() {
         short producerEpoch = 0;
 
-        ProducerStateManager.ProducerAppendInfo appendInfo = stateManager.prepareUpdate(
+        ProducerAppendInfo appendInfo = stateManager.prepareUpdate(
                 producerId, PartitionLog.AppendOrigin.Client);
         appendInfo.appendDataBatch(producerEpoch, 0, 5, time.milliseconds(),
                 15L, 20L);
@@ -385,13 +383,13 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         stateManager.update(appendInfo);
         assertTrue(stateManager.lastEntry(producerId).isPresent());
 
-        ProducerStateManager.ProducerAppendInfo nextAppendInfo = stateManager.prepareUpdate(
+        ProducerAppendInfo nextAppendInfo = stateManager.prepareUpdate(
                 producerId, PartitionLog.AppendOrigin.Client);
         nextAppendInfo.appendDataBatch(producerEpoch, 6, 10, time.milliseconds(),
                 26L, 30L);
         assertTrue(stateManager.lastEntry(producerId).isPresent());
 
-        ProducerStateManager.ProducerStateEntry lastEntry = stateManager.lastEntry(producerId).get();
+        ProducerStateEntry lastEntry = stateManager.lastEntry(producerId).get();
         assertEquals(0, lastEntry.firstSeq().intValue());
         assertEquals(5, lastEntry.lastSeq().intValue());
         assertEquals(20L, lastEntry.lastDataOffset().longValue());
@@ -410,21 +408,21 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         long offset = 9L;
         append(stateManager, producerId, producerEpoch, 0, offset);
 
-        ProducerStateManager.ProducerAppendInfo appendInfo =
+        ProducerAppendInfo appendInfo =
                 stateManager.prepareUpdate(producerId, PartitionLog.AppendOrigin.Client);
         appendInfo.appendDataBatch(producerEpoch, 1, 5, time.milliseconds(),
                 16L, 20L);
         // Update to real offset
         appendInfo.updateCurrentTxnFirstOffset(true, 16L);
-        ProducerStateManager.ProducerStateEntry lastEntry = appendInfo.toEntry();
-        assertEquals(producerEpoch, lastEntry.getProducerEpoch().shortValue());
+        ProducerStateEntry lastEntry = appendInfo.toEntry();
+        assertEquals(producerEpoch, lastEntry.producerEpoch().shortValue());
         assertEquals(1, lastEntry.firstSeq().intValue());
         assertEquals(5, lastEntry.lastSeq().intValue());
         assertEquals(16L, lastEntry.firstDataOffset().longValue());
         assertEquals(20L, lastEntry.lastDataOffset().longValue());
-        assertEquals(Optional.of(16L), lastEntry.getCurrentTxnFirstOffset());
+        assertEquals(Optional.of(16L), lastEntry.currentTxnFirstOffset());
         assertEquals(
-                Lists.newArrayList(new ProducerStateManager.TxnMetadata(producerId, 16L)),
+                Lists.newArrayList(new TxnMetadata(producerId, 16L)),
                 appendInfo.startedTransactions());
 
         appendInfo.appendDataBatch(producerEpoch, 6, 10, time.milliseconds(),
@@ -432,33 +430,33 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         // Update to real offset
         appendInfo.updateCurrentTxnFirstOffset(true, 26L);
         lastEntry = appendInfo.toEntry();
-        assertEquals(producerEpoch, lastEntry.getProducerEpoch().shortValue());
+        assertEquals(producerEpoch, lastEntry.producerEpoch().shortValue());
         assertEquals(6, lastEntry.firstSeq().intValue());
         assertEquals(10, lastEntry.lastSeq().intValue());
-        assertEquals(Optional.of(16L), lastEntry.getCurrentTxnFirstOffset());
+        assertEquals(Optional.of(16L), lastEntry.currentTxnFirstOffset());
         assertEquals(
-                Lists.newArrayList(new ProducerStateManager.TxnMetadata(producerId, 16L)),
+                Lists.newArrayList(new TxnMetadata(producerId, 16L)),
                 appendInfo.startedTransactions());
 
         EndTransactionMarker endTxnMarker = new EndTransactionMarker(ControlRecordType.COMMIT, coordinatorEpoch);
-        Optional<PartitionLog.CompletedTxn> completedTxnOpt =
+        Optional<CompletedTxn> completedTxnOpt =
                 appendInfo.appendEndTxnMarker(endTxnMarker, producerEpoch, 40L, time.milliseconds());
         assertTrue(completedTxnOpt.isPresent());
 
-        PartitionLog.CompletedTxn completedTxn = completedTxnOpt.get();
-        assertEquals(producerId, completedTxn.getProducerId());
-        assertEquals(16L, completedTxn.getFirstOffset().longValue());
-        assertEquals(40L, completedTxn.getLastOffset().longValue());
-        assertFalse(completedTxn.getIsAborted());
+        CompletedTxn completedTxn = completedTxnOpt.get();
+        assertEquals(producerId, completedTxn.producerId());
+        assertEquals(16L, completedTxn.firstOffset().longValue());
+        assertEquals(40L, completedTxn.lastOffset().longValue());
+        assertFalse(completedTxn.isAborted());
 
         lastEntry = appendInfo.toEntry();
-        assertEquals(producerEpoch, lastEntry.getProducerEpoch().shortValue());
+        assertEquals(producerEpoch, lastEntry.producerEpoch().shortValue());
         // verify that appending the transaction marker doesn't affect the metadata of the cached record batches.
         assertEquals(6, lastEntry.firstSeq().intValue());
         assertEquals(10, lastEntry.lastSeq().intValue());
-        assertEquals(Optional.empty(), lastEntry.getCurrentTxnFirstOffset());
+        assertEquals(Optional.empty(), lastEntry.currentTxnFirstOffset());
         assertEquals(
-                Lists.newArrayList(new ProducerStateManager.TxnMetadata(producerId, 16L)),
+                Lists.newArrayList(new TxnMetadata(producerId, 16L)),
                 appendInfo.startedTransactions());
     }
 
@@ -490,7 +488,7 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
 
         // Append with the bumped epoch should be fine if starting from sequence 0
         append(stateManager, producerId, bumpedEpoch, 0, 0L, time.milliseconds(), true);
-        assertEquals(bumpedEpoch, stateManager.lastEntry(producerId).get().getProducerEpoch().shortValue());
+        assertEquals(bumpedEpoch, stateManager.lastEntry(producerId).get().producerEpoch().shortValue());
         assertEquals(0, stateManager.lastEntry(producerId).get().lastSeq().shortValue());
     }
 
@@ -508,8 +506,8 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         appendEndTxnMarker(stateManager, producerId, (short) (epoch + 1), ControlRecordType.ABORT,
                 1L, -1, time.milliseconds());
 
-        ProducerStateManager.ProducerStateEntry lastEntry = stateManager.lastEntry(producerId).get();
-        assertEquals(Optional.empty(), lastEntry.getCurrentTxnFirstOffset());
+        ProducerStateEntry lastEntry = stateManager.lastEntry(producerId).get();
+        assertEquals(Optional.empty(), lastEntry.currentTxnFirstOffset());
         assertEquals(-1, lastEntry.lastDataOffset().longValue());
         assertEquals(-1, lastEntry.firstDataOffset().longValue());
 
@@ -595,7 +593,7 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
 
         // Appending the empty control batch should not throw and a new transaction shouldn't be started
         append(stateManager, producerId, baseOffset, batch, PartitionLog.AppendOrigin.Client);
-        assertEquals(Optional.empty(), stateManager.lastEntry(producerId).get().getCurrentTxnFirstOffset());
+        assertEquals(Optional.empty(), stateManager.lastEntry(producerId).get().currentTxnFirstOffset());
     }
 
     @Test(timeOut = defaultTestTimeout)
@@ -638,17 +636,17 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
         latch.await();
     }
 
-    private Optional<PartitionLog.CompletedTxn> appendEndTxnMarker(ProducerStateManager mapping,
+    private Optional<CompletedTxn> appendEndTxnMarker(ProducerStateManager mapping,
                                                                            Long producerId,
                                                                            Short producerEpoch,
                                                                            ControlRecordType controlType,
                                                                            Long offset,
                                                                            Integer coordinatorEpoch,
                                                                            Long timestamp) {
-        ProducerStateManager.ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(
+        ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(
                 producerId, PartitionLog.AppendOrigin.Coordinator);
         EndTransactionMarker endTxnMarker = new EndTransactionMarker(controlType, coordinatorEpoch);
-        Optional<PartitionLog.CompletedTxn> completedTxnOpt =
+        Optional<CompletedTxn> completedTxnOpt =
                 producerAppendInfo.appendEndTxnMarker(endTxnMarker, producerEpoch, offset, timestamp);
         mapping.update(producerAppendInfo);
         completedTxnOpt.ifPresent(mapping::completeTxn);
@@ -684,7 +682,7 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
                         Long timestamp,
                         Boolean isTransactional,
                         PartitionLog.AppendOrigin origin) {
-        ProducerStateManager.ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(producerId, origin);
+        ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(producerId, origin);
         producerAppendInfo
                 .appendDataBatch(producerEpoch, seq, seq, timestamp, -1L, -1L);
         // Update to real offset
@@ -698,7 +696,7 @@ public class ProducerStateManagerTest extends KopProtocolHandlerTestBase {
                         Long offset,
                         RecordBatch batch,
                         PartitionLog.AppendOrigin origin) {
-        ProducerStateManager.ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(producerId, origin);
+        ProducerAppendInfo producerAppendInfo = stateManager.prepareUpdate(producerId, origin);
         producerAppendInfo.append(batch, Optional.empty());
         stateManager.update(producerAppendInfo);
         stateManager.updateMapEndOffset(offset + 1);
